LocationDetail을 가져오는 코드 작성
LocationDetail을 가져오는 코드 작성

+++ Backend/bfSearcher/src/main/java/softwareEngineering/bfSearcher/Repository/ReviewRepository.java
@@ -9,5 +9,5 @@ public interface ReviewRepository extends JpaRepository<Review,Long> {
 
     Review save(Review review);
 
-    List<Review> findByLocationId(int location_id);
+    List<Review> findByLocationId(Long location_id);
 }
